requirements.txt
selenium
requests
allure-python-commons
python-dotenv